Item are able to be used.. Juz need to implement into BattleScene

+++ SP4/Base/Source/Main/Game/Objects/Items/Consumables.h
@@ -14,7 +14,7 @@ public:
 	int UseHealthPot(int HP);
 	int UseManaPot(int Mana);
 	//int UseBandage()
-	int UseAtkBuffPot(int Attack);
-	int UseDefBuffPot(int Defense);
+	float UseAtkBuffPot(int Attack);
+	float UseDefBuffPot(int Defense);
 };
